Rename types.d.ts to type.ts
Declaration files should only be used to represent types that result
from the transpilation of TypeScript to JavaScript. In a TypeScript
context, it can cause issues that, from my understanding, relate to
ambient types.
