+++ js/loader.ts
@@ -1,6 +1,6 @@
 // Copyright 2018-2023 the Deno authors. All rights reserved. MIT license.
 
-import type { LoadResponse } from "./types.d.ts";
+import type { LoadResponse } from "./types.ts";
 
 const hasPermissions = "permissions" in Deno;
 let readRequested = false;
